Refactor since there was a unnecessary folder
