+++ lib/connectMongoose.js
