+++ package.json
@@ -1,4 +1,13 @@
 {
+  "name": "flitter-backend",
+  "version": "0.0.0",
+  "private": true,
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node ./bin/www",
+    "dev": "cross-env DEBUG=nodepop:* PORT=3000 nodemon ./bin/www",
+    "init-db": "node init-db.js"
+  },
   "dependencies": {
     "cookie-parser": "~1.4.4",
     "cross-env": "^7.0.3",
@@ -11,17 +20,5 @@
     "mongoose": "^6.8.2",
     "morgan": "~1.9.1",
     "nodemon": "^2.0.20"
-  },
-  "name": "flitter-backend",
-  "version": "1.0.0",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node ./bin/www",
-    "dev": "cross-env DEBUG=nodepop:* PORT=3000 nodemon ./bin/www",
-    "init-db": "node init-db.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "description": ""
+  }
 }
